DRY up the app-specific DB admin box configuration
We're about to create many app-specific DB admin boxes, all of
which will have the same configuration on the node itself, so it
makes sense to encapsulate this in one place.

I did consider leaving the existing db_admin node untouched, as
we will eventually remove the Postgres code from it when every
Postgres-powered app is using its own RDS instance. However, it
uses the same code we've just encapsulated (albeit in a slightly
different order), so there's little harm in making this optimisation
now.
I did have to move the "include all PostgreSQL classes that create
databases and users" declaration higher up so that all Postgres
config was in the same place, and I avoided putting the Postgres
config at the end of the file in case this class relies on these
lines being at the beginning (which are now pulled in via the class).
```
  include govuk_env_sync
  include ::govuk::node::s_base
```